Fixe Update Movie Image

+++ middleware/imageValidation.js
@@ -1,29 +1,33 @@
 const fs = require('fs');
 const AppError = require('../utils/AppError');
 
-module.exports =  async(req,resp, next) => {
+module.exports = async (req, resp, next) => {
 
-    let name = req.file.filename;
-    let image = req.file.path
+    let isImageExists = false;
 
-    if(!(req.file.mimetype).includes('jpg')
-        && !(req.file.mimetype).includes('png')
-        && !(req.file.mimetype).includes('bmp')
-        && !(req.file.mimetype).includes('jpeg')){
-        fs.unlinkSync(image);
+    if (req.file && req.filename && req.file.path) {
 
-        return next(new AppError('File Type is not supported', 400));
+        let name = req.file.filename;
+        let image = req.file.path
+
+        if (!(req.file.mimetype).includes('jpg')
+            && !(req.file.mimetype).includes('png')
+            && !(req.file.mimetype).includes('bmp')
+            && !(req.file.mimetype).includes('jpeg')) {
+            fs.unlinkSync(image);
+
+            return next(new AppError('File Type is not supported', 400));
+        }
+
+        if (req.file.size > 1024 * 1024) {
+            fs.unlinkSync(image);
+            return next(new AppError('File is too large', 400));
+        }
+
+        isImageExists = true;
     }
 
-    if(req.file.size > 1024 * 1024){
-        fs.unlinkSync(image);
-        return next(new AppError('File is too large', 400));
-    }
-
-    if(!name || !image){
-        return next(new AppError('Image is required', 400));
-    }
-
+    resp.locals.isImageExists = isImageExists;
 
     next();
 }
